add a new IT class
